Make handlers a little less gross [#12]
A preliminary stab at improving the end-user interface for defining
command and file handlers. We've ditched the manually defined
dictionaries in favour of base classes: `CommandHandler` and `FiletypeHandler`.

Handlers have been moved to `handlers.py`, away from `cmd.py` which now
solely provides the functionality that dispatches the relevant handlers
if required. This is a work in progress.